+++ src/main/java/org/fenixedu/academictreasury/domain/reports/DebtReportRequestResultFile.java
@@ -39,10 +39,15 @@ package org.fenixedu.academictreasury.domain.reports;
 
 import static org.fenixedu.academictreasury.util.AcademicTreasuryConstants.academicTreasuryBundle;
 
+import java.io.InputStream;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.stream.Stream;
 
+import com.qubit.terra.framework.services.ServiceProvider;
+import com.qubit.terra.framework.services.fileSupport.FileDescriptor;
+import com.qubit.terra.framework.services.fileSupport.FileManager;
+import org.apache.commons.lang.StringUtils;
 import org.fenixedu.academictreasury.domain.exceptions.AcademicTreasuryDomainException;
 import org.fenixedu.bennu.io.domain.IGenericFile;
 import org.fenixedu.treasury.services.accesscontrol.TreasuryAccessControlAPI;
@@ -75,6 +80,11 @@ public class DebtReportRequestResultFile extends DebtReportRequestResultFile_Bas
         final ITreasuryPlatformDependentServices services = TreasuryPlataformDependentServicesFactory.implementation();
         
         services.createFile(this, filename, CONTENT_TYPE, content);
+
+        final FileManager fileManager = ServiceProvider.getService(FileManager.class);
+
+        FileDescriptor fileDescriptor = fileManager.createFile(filename, content.length, CONTENT_TYPE, content);
+        setFileDescriptorId(fileDescriptor.getId());
 
         checkRules();
     }
@@ -114,12 +124,20 @@ public class DebtReportRequestResultFile extends DebtReportRequestResultFile_Bas
     
     @Override
     public void delete() {
-        
         setDomainRoot(null);
+        setDebtReportRequest(null);
         
         final ITreasuryPlatformDependentServices services = TreasuryPlataformDependentServicesFactory.implementation();
-        services.deleteFile(this);
-        
+        FileManager fileManager = ServiceProvider.getService(FileManager.class);
+
+        if(StringUtils.isNotEmpty(getFileDescriptorId())) {
+            fileManager.delete(getFileDescriptorId());
+        }
+
+        if (getTreasuryFile() != null) {
+            services.deleteFile(this);
+        }
+
         super.deleteDomainObject();
     }
 
@@ -138,5 +156,64 @@ public class DebtReportRequestResultFile extends DebtReportRequestResultFile_Bas
     public static Stream<DebtReportRequestResultFile> findAll() {
         return FenixFramework.getDomainRoot().getDebtReportRequestResultFilesSet().stream();
     }
-    
+
+    @Override
+    public byte[] getContent() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getContent();
+        }
+
+        return IGenericFile.super.getContent();
+    }
+
+    @Override
+    public long getSize() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getSize();
+        }
+
+        return IGenericFile.super.getSize();
+    }
+
+    @Override
+    public String getFilename() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getName();
+        }
+
+        return IGenericFile.super.getFilename();
+    }
+
+    @Override
+    public String getContentType() {
+        FileDescriptor fd = getFileDescriptor();
+        if (fd != null) {
+            return fd.getContentType();
+        }
+
+        return IGenericFile.super.getContentType();
+    }
+
+    @Override
+    public InputStream getStream() {
+        FileDescriptor fd = getFileDescriptor();
+
+        if (fd != null) {
+            return fd.getReadStream();
+        }
+
+        return IGenericFile.super.getStream();
+    }
+
+    private FileDescriptor getFileDescriptor() {
+        if (StringUtils.isNotBlank(getFileDescriptorId())) {
+            return ServiceProvider.getService(FileManager.class).getFileDescriptor(getFileDescriptorId());
+        }
+
+        return null;
+    }
+
 }
